+++ Loader/AbstractCryptoLoader.php
@@ -1,0 +1,19 @@
+<?php
+
+/*
+ * This file is part of the MesCryptoBundle package.
+ *
+ * (c) Francesco Cartenì <http://www.multimediaexperiencestudio.it/>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Mes\Security\CryptoBundle\Loader;
+
+/**
+ * Class AbstractCryptoLoader.
+ */
+abstract class AbstractCryptoLoader implements CryptoLoaderInterface
+{
+}